git update xserver libxcb libxcb/xcb-proto xkeyboard-config mkfontscale libxtrans mesa plink until 21 June 2016
xserver          commit 111a045dcf0912dc5b178c18acda479b3972876b
libxcb           commit 65b298c7ca317d7e4316aa2b9e0499e13047c65c
libxcb/xcb-proto commit 95a262e0e66cd88b9d7a133917b3ba70ace77301
xkeyboard-config commit deb6293f48bd81e5a4110d09a0b30caf8e2a28d1
mkfontscale      commit ecb248d8865df376a97a795c2e4f58a3bf64c3f1
libxtrans        commit 3810e026323210df6ec6cf818f01832a63e9fbb9
mesa             commit b16d21270f9ea0173ebe059a3e82a4b451a80596
putty            commit 7b9ad090067e817321190fa3379264e6ee9d3b91

+++ mesalib/src/compiler/glsl/ast_array_index.cpp
@@ -92,12 +92,12 @@ update_max_array_access(ir_rvalue *ir, int idx, YYLTYPE *loc,
                deref_record->record->type->field_index(deref_record->field);
             assert(field_index < deref_var->var->get_interface_type()->length);
 
-            unsigned *const max_ifc_array_access =
+            int *const max_ifc_array_access =
                deref_var->var->get_max_ifc_array_access();
 
             assert(max_ifc_array_access != NULL);
 
-            if (idx > (int)max_ifc_array_access[field_index]) {
+            if (idx > max_ifc_array_access[field_index]) {
                max_ifc_array_access[field_index] = idx;
 
                /* Check whether this access will, as a side effect, implicitly
